notebooks/I&W Reporting/Production Scripts/I&W_Document_Generator/requirements.txt
# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# API Requests and SSL Handling
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.12.7

# Configuration Management
python-dotenv>=0.21.0

# Report Generation
python-docx>=0.8.11

# Logging and Console Output Formatting
rich>=13.0.0

# Testing
pytest>=7.2.0
unittest2>=1.1.0

# JSON Handling and Data Normalization
jsonschema>=4.16.0

# Datetime Handling and Timezone Support
pytz>=2022.7.0

# Optional: Progress Bar for Fetching Data
tqdm>=4.64.0
